refactor: Use the logic to get access to the NuGet packages in the build output in BuildVerification tests in MSBuild integration tests
Move logic to load packages from BuildVerification project to TestHelpers and reuse it in MSBuild.IntegrationTest.
This allows removing the custom build target from the MSBuild integration tests project and avoids generating the NuGet packages multiple times during build.
However, this also means the MSBuild integration tests can only be started via the build.ps1 script and will no longer work when started directly from dotnet test or the Visual Studio Test Explorer.
