+++ src/MdDocs.TestHelpers/Grynwald.MdDocs.TestHelpers.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="Mono.Cecil" Version="0.11.4" />
     <PackageReference Include="ApprovalTests" Version="5.8.0" />
     <PackageReference Include="Basic.Reference.Assemblies" Version="1.4.1" />
+    <PackageReference Include="NuGet.Packaging" Version="6.4.0" />
   </ItemGroup>
 
 </Project>
